+++ packages/view_model_generator/example/lib/example.vm.dart
@@ -1,5 +1,5 @@
-// GENERATED CODE - DO NOT MODIFY BY HAND
 // dart format width=80
+// GENERATED CODE - DO NOT MODIFY BY HAND
 
 part of 'example.dart';
 
@@ -13,4 +13,11 @@ final counterViewModelProvider = ViewModelProvider<CounterViewModel>(
 
 final userViewModelProvider = ViewModelProvider.arg<UserViewModel, Repository>(
   builder: (Repository repo) => UserViewModel(repo),
+);
+
+final userKeyViewModelProvider =
+    ViewModelProvider.arg<UserKeyViewModel, Repository>(
+  builder: (Repository repo) => UserKeyViewModel(repo),
+  key: (Repository repo) => repo,
+  tag: (Repository repo) => 'user_key',
 );